Unroll loops

+++ src/ImageSharp/Common/Helpers/SimdUtils.Shuffle.cs
@@ -229,7 +229,7 @@ namespace SixLabors.ImageSharp
             public const byte ZYXW = (3 << 6) | (0 << 4) | (1 << 2) | 2;
 
             [MethodImpl(InliningOptions.ShortMethod)]
-            public static byte MmShuffle(int p3, int p2, int p1, int p0)
+            public static byte MmShuffle(byte p3, byte p2, byte p1, byte p0)
                 => (byte)((p3 << 6) | (p2 << 4) | (p1 << 2) | p0);
 
             [MethodImpl(InliningOptions.ShortMethod)]
